Built-in types now are very similar with user-defined types. They have the same representation, common access macros to get the name and are all stored in type_list.
Expands and functions are stored in a function_list variable.
After parsing we chek if there were no parsing errors and
only then are going to perform typecheking.

Printing routines now are fully implemented. There is a top-level
function print_all (FILE *), that prints all the program. However,
if you want to print a specific part of the program like function
or expression, there is a set of functions for this purpose as well.

strlist construction can be parsed successfully and it is stored
in the constant_list variable.

Should be it for the commit. :)