Typechecking is almost finished.
Typechecker can implicitly insert type-conversions if they are in
the function prototype list. For statements are not implemented (yet)
the implementation there should be trivial.

The problem of type list element still exists. I don't know a
good solution. The plan is introduce a new type LIST_ELEMENT or
ANY which basically could be anything.

All the rest works fine.