User defined prototypes with ccall are intoduced.
Now user can define a prototype for the external function specifying
the name of the function that would be code-generated. For example
if you wan to use function 'get' at the level of expand, which would
be translated into the C function call GetX, you can write the
following prototype:

proto name="get" ret=int args=[int x] ccall="GetX";

This will make the typechecker happy and the codege-generator will
know what to generate. Also this mechanism allows you to redefine
existing code-generation for the built-in types.